[BUG] Hot fix for issue #97
SASS compilation changed the class name order for the btc modal. This
interferred with the re-usable toggle.js code which grabs on to the
first class name. The --active modifier class is now
applied to class "footer__btc__modal" instead of "modal". Hence, the
modal didn't open anymore in the production build. Instead of
messing with the JS code or diving into SASS configurations, this
hot fix simply duplicates the --active class css code in the
"footer__btc__modal" class. Now the modal does open as expected.